:art: Move docker api classes to docker package
